생체데이터 업로드 시 GSR파일 데이터만 업로드 되는 문제 수정 ECG파일 파싱 index 수정

+++ web/src/main/java/com/stp/absm/controller/Page002Controller.java
@@ -224,8 +224,9 @@ public class Page002Controller extends RootController {
                 param.put("prId",prId);
 
                 if ("ECG".equals(fileType[i]) || "GSR".equals(fileType[i])) {
-                    //측정 데이터삭제
-                    page002Mapper.deleteMeasure(param);
+                    //측정 데이터삭제 ECG일 경우만 삭제
+                    if ("ECG".equals(fileType[i]))
+                        page002Mapper.deleteMeasure(param);
 
                     measureFileService.setFileInfo(fileUploadInfo);
                     measureFileService.doParse();
